+++ packages/validation-plugin/src/validators/options.ts
@@ -12,6 +12,7 @@ const VALID_QUERY_PARAMS = {
   arrayFormat: ["none", "brackets", "index", "comma"] as const,
   booleanFormat: ["none", "string", "empty-true"] as const,
   nullFormat: ["default", "hidden"] as const,
+  numberFormat: ["none", "auto"] as const,
 } as const;
 
 const VALID_LOGGER_LEVELS = [